// fron/cliente/src/app/models/Edm05.ts
export class Edm05casillas{
    _id?: number;
    edm05casilla1: string;
    edm05casilla2: string;
    edm05casilla3: string;
    edm05casilla4: string;
    edm05casilla5: string;
    edm05casilla6: string;
    edm05casilla7: string;
    edm05casilla8: string;
    edm05casilla9: string;
    edm05casilla10: string;
    edm05casilla11: string;
    edm05casilla12: string;
    edm05casilla13: string;
    edm05casilla14: string;
    edm05casilla15: string;
    edm05casilla16:string;
    edm05casilla17: string;
    edm05casilla18: string;
    edm05casilla19: string;
    edm05casilla20: string;
    edm05casilla21: string;
    edm05casilla22: string;
    edm05casilla23: string;
    edm05casilla24: string;
    edm05casilla25: string;
    edm05casilla26: string;
    edm05casilla27: string;
    

    constructor(edm05casilla1:string, edm05casilla2:string,edm05casilla3:string,edm05casilla4:string,edm05casilla5:string ,edm05casilla6: string,
       edm05casilla7: string,edm05casilla8: string,edm05casilla9: string, edm05casilla10: string,edm05casilla11: string,edm05casilla12: string,edm05casilla13: string,edm05casilla14: string,
        edm05casilla15: string,edm05casilla16: string,edm05casilla17: string,edm05casilla18: string,edm05casilla19: string,edm05casilla20: string,edm05casilla21: string,
        edm05casilla22: string,edm05casilla23: string, edm05casilla24: string,edm05casilla25: string,edm05casilla26: string,edm05casilla27: string,
       ){
        this.edm05casilla1=edm05casilla1;
        this.edm05casilla2=edm05casilla2;
        this.edm05casilla3=edm05casilla3;
        this.edm05casilla4=edm05casilla4;
        this.edm05casilla5=edm05casilla5;
        this.edm05casilla6=edm05casilla6;
        this.edm05casilla7=edm05casilla7;
        this.edm05casilla8=edm05casilla8;
        this.edm05casilla9=edm05casilla9;
        this.edm05casilla10=edm05casilla10;
        this.edm05casilla11=edm05casilla11;
        this.edm05casilla12=edm05casilla12;
        this.edm05casilla13=edm05casilla13;
        this.edm05casilla14=edm05casilla14;
        this.edm05casilla15=edm05casilla15;
        this.edm05casilla16=edm05casilla16;
        this.edm05casilla17=edm05casilla17;
        this.edm05casilla18=edm05casilla18;
        this.edm05casilla19=edm05casilla19;
        this.edm05casilla20=edm05casilla20;
        this.edm05casilla21=edm05casilla21;
        this.edm05casilla22=edm05casilla22;
        this.edm05casilla23=edm05casilla23;
        this.edm05casilla24=edm05casilla24;
        this.edm05casilla25=edm05casilla25;
        this.edm05casilla26=edm05casilla26;
        this.edm05casilla27=edm05casilla27;

    }

}